R6 Class representing a K Armed Bandit Agent

R6 Class representing a K Armed Bandit Agent

R6 Class representing a K Armed Bandit Agent

Details

Called by rl_new_agent.

Public fields

model_id

(character) A model identifier referencing the reinforcement learning paradigm to perform the simulation.

num_trials

(numeric) The number of trials to simulate.

num_episodes

(numeric) The number of episodes per trial.

num_arms

(numeric) The number of arms (options) the agent can sample from.

gamma

(numeric) The temporal discounting factor of the RL agent

alpha

(numeric) The learning rate of the RL agent

action_episode

(numeric) The episode number in which the agent takes an action.

reinforcement_episode

(numeric) The episode number in which the agent receives the reinforcement after taking an action.

policy

(list) A list containing the information relevant to the agent's decision policy, such as 'softmax' or 'epsilon-greedy' that guides choice behavior

actions

A numeric vector containing the index of which arm was selected on a given trial.

arm_structure

A list of arm definitions where each element contains a data frame with columns 'probability', 'magnitude', 'alternative', and 'trial' describing, respectively, the probability of receiving a reward magnitude with the alternative for each trial.

Q_values

A three dimensional array tracking the estimated Q-value for selecting each arm across all episodes and trials

prediction_errors

A three dimensional array tracking the prediction errors associated with choosing each arm across all episodes and trials

reinforcements

A matrix tracking the experienced reinforcements associated with choosing an arm across episodes within a given trial

Methods

Public methods

agent_k_armed_bandit$new()

agent_k_armed_bandit$print()

agent_k_armed_bandit$set_policy()

agent_k_armed_bandit$set_arms()

agent_k_armed_bandit$get_arms()

agent_k_armed_bandit$simulate_agent()

agent_k_armed_bandit$get_learned_values()

agent_k_armed_bandit$get_pe_data()

agent_k_armed_bandit$clone()

Method new()

Create a new kArmedBandit object

Usage

agent_k_armed_bandit$new(
  model_id,
  num_trials,
  num_episodes,
  num_arms,
  action_episode,
  reinforcement_episode,
  gamma,
  alpha
)

Arguments

model_id

model_id (character) A model identifier referencing the reinforcement learning paradigm to perform the simulation.

num_trials

(numeric) The number of trials to simulate.

num_episodes

(numeric) The number of episodes per trial.

num_arms

(numeric) The number of arms (options) the agent can sample from.

action_episode

(numeric) The episode an action should be taken on.

reinforcement_episode

(numeric) The episode reinforcements should occur on.

gamma

(numeric) The temporal discounting factor of the RL agent

alpha

(numeric) The learning rate of the RL agent

Method print()

Printing method for object of class 'rlAgent'.

Usage

agent_k_armed_bandit$print(...)

Arguments

...

NA; printing function

Method set_policy()

Set an Agent's Action-Selection Policy

Usage

agent_k_armed_bandit$set_policy(policy, ...)

Arguments

policy

What policy should a decision be made under? Currently supported are softmax, greedy, and epsilon-greedy.

...

Additional arguments passed to or from specific methods, such as tau when policy = "softmax" and epsilon when policy = "epsilonGreedy".

Method set_arms()

Define the arm structure for an agent

Usage

agent_k_armed_bandit$set_arms(arm_input, keep_arm_structure = FALSE)

Arguments

arm_input

A list of arm definitions where each element contains a data frame with columns 'probability', 'magnitude', 'alternative', and 'trial' describing, respectively, the probability of receiving a reward magnitude with the alternative for each trial.

keep_arm_structure

(Logical) FALSE (default) and any existing arm structure will be replaced when called. TRUE and the arm structure will be modified but will not remove previously defined rewards.

Method get_arms()

Retrieve the list of arm definitions

Usage

agent_k_armed_bandit$get_arms()

Returns

A list of data frames

Method simulate_agent()

Simulate the RL Agent

Usage

agent_k_armed_bandit$simulate_agent()

Method get_learned_values()

Convert the agent's estimated Q values array where, for each action, there is an matrix where each row is an episode and each column is a trial that contains the q_value, to a dataframe with columns 'trial', 'episode', 'action', and 'q_value'.

Usage

agent_k_armed_bandit$get_learned_values(
  add_trial_zero = TRUE,
  trial_zero_value = 0
)

Arguments

add_trial_zero

(Logical) TRUE by default and trial zero will be appended to the prediction error data frame with values from trial_zero_value. FALSE and output will begin at trial one.

trial_zero_value

(Numeric) Either a single value (default is 0) or a vector of values with length num_arms times num_episodes to append for trial 0.

Returns

A dataframe with the agent's simulated learned values ('q_value') from taking an action for each episode across all trials.

Method get_pe_data()

Convert the agent's prediction error array where, for each action, there is an matrix with each row corresponding to an episode and each column a trial with the experienced prediction error, to a dataframe with columns 'trial', 'episode', 'action', and 'value'.

Usage

agent_k_armed_bandit$get_pe_data(add_trial_zero = TRUE, trial_zero_value = 0)

Arguments

add_trial_zero

(Logical) TRUE by default and trial zero will be appended to the prediction error data frame with values from trial_zero_value. FALSE and output will begin at trial one.

trial_zero_value

(Numeric) Either a single value (default is 0) or a vector of values with length num_arms times num_episodes to append for trial 0.

Returns

A dataframe with the agent's simulated prediction errors from taking an action for each episode across all trials.

Method clone()

The objects of this class are cloneable with this method.

Usage

agent_k_armed_bandit$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.