R6 Class representing a new Reinforcement Learning Agent

R6 Class representing a new Reinforcement Learning Agent

R6 Class representing a new Reinforcement Learning Agent

Details

Called by rl_new_agent.

Public fields

model_type

(character) either "tdrl" or "vprl", referencing the reinforcement learning paradigm to perform the simulation.

num_trials

(numeric) The number of trials to simulate.

num_episodes

(numeric) The number of episodes per trial.

num_cues

(numeric) The number of potential cues interacting with the RL agent.

gamma

(numeric) The temporal discounting factor of the RL agent

alpha

(numeric) The learning rate of the RL agent

present_cues

A three-dimensional array tacking what cues are present in each episode across all simulated trials

estimated_value

A three dimensional array tracking the estimated values of each episode across all trials

reward

A matrix tracking experienced rewards in each episode across all trials

RPE

A matrix tracking the reward prediction errors occuring in

Methods

Public methods

rl_agent$new()

rl_agent$print()

rl_agent$set_reward()

rl_agent$set_cues()

rl_agent$simulate_agent()

rl_agent$get_tidy_pe_data()

rl_agent$clone()

Method new()

Create a new rlAgent object

Usage

rl_agent$new(model_type, num_trials, num_episodes, num_cues, gamma, alpha)

Arguments

model_type

(character) either "tdrl" or "vprl", referencing the reinforcement learning paradigm to perform the simulation.

num_trials

(numeric) The number of trials to simulate.

num_episodes

(numeric) The number of episodes per trial.

num_cues

(numeric) The number of potential cues interacting with the RL agent.

gamma

(numeric) The temporal discounting factor of the RL agent

alpha

(numeric) The learning rate of the RL agent

Method print()

Printing method for object of class 'rlAgent'.

Usage

rl_agent$print(...)

Arguments

...

NA; printing function

Method set_reward()

Define the onset episode and offset episode of rewards for each trial

Usage

rl_agent$set_reward(
  reward_onset,
  reward_offset,
  reward_magnitude,
  keep_reward_structure = FALSE
)

Arguments

reward_onset

(Numeric) The episode number a reward presentation begins. Either a single number where reward is presented at the same episode in all simulated trial or a numeric vector of length num_trials with reward onset occurring in different episodes.

reward_offset

(Numeric) The episode number a reward presentation ends. Either a single number where reward is presented at the same episode in all simulated trial or a numeric vector of length num_trials with reward offset occurring in different episodes.

reward_magnitude

(Numeric) The magnitude of the reward. Either a single number where reward is presented with the same magnitude across all simulated trials or a numeric vector of length num_trials with reward magnitude differing across trials.

keep_reward_structure

(Logical) FALSE (default) and any existing reward structure will be replaced when called. TRUE and the reward structure will be modified but will not remove previously defined rewards.

Method set_cues()

Define the onset episode and offset episode of cues for each trial

Usage

rl_agent$set_cues(cue_list, keep_cue_structure = FALSE)

Arguments

cue_list

A list of length num_cues where each element contains a data frame with columns 'cue', 'onset', 'offset', and 'trial'.

keep_cue_structure

(Logical) FALSE (default) and any existing cue structure will be replaced when called. TRUE and the cue structure will be modified but will not remove previously defined cues.

Examples

cue_data_list <- list(
  one = data.frame(
    cue = 1,
    onset = 60,
    offset = 80,
    trial = 1:500),
  two = data.frame(
    cue = 2,
    onset = 70,
    offset = 80,
    trial = 1:500
  )
)

Method simulate_agent()

Simulate the RL Agent

Usage

rl_agent$simulate_agent()

Method get_tidy_pe_data()

Convert the agent's simulated reward prediction errors from a matrix where each row is an episode and each column is a trial to a dataframe with columns 'trial', 'episode', and 'value'.

Usage

rl_agent$get_tidy_pe_data(add_trial_zero = TRUE, trial_zero_value = 0)

Arguments

add_trial_zero

(Logical) TRUE by default and trial zero will be appended to the prediction error data frame with values from trial_zero_value. FALSE and output will begin at trial one.

trial_zero_value

(Numeric) Either a single value (default is 0) or a vector of values to append for trial 0.

Returns

A dataframe with the agent's simulated reward prediction errors ('value') for each episode across trials.

Method clone()

The objects of this class are cloneable with this method.

Usage

rl_agent$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## ------------------------------------------------
## Method `rl_agent$set_cues`
## ------------------------------------------------

cue_data_list <- list(
  one = data.frame(
    cue = 1,
    onset = 60,
    offset = 80,
    trial = 1:500),
  two = data.frame(
    cue = 2,
    onset = 70,
    offset = 80,
    trial = 1:500
  )
)