Define an RL Arm

Define the probabilistic-reinforcement structure of a specific arm. This function is a simple wrapper around data.frame. As such, you can pass in either numeric vectors the same length as the number of simulated trials or a single number.

rl_arms_define(probability, magnitude, alternative, trial)

Arguments

probability

Either a numeric vector the same length as the number of simulated trials or a single number between zero and one to define the probability a reinforcement will be delivered.

magnitude

Either a numeric vector the same length as the number of simulated trials or a single number defining the magnitude of the reinforcement to be delivered with probability probability if this arm is chosen.

alternative

Either a numeric vector the same length as the number of simulated trials or a single number defining the magnitude of the reinforcement to be delivered with probability (1 - probability) if this arm is chosen.

trial

A numeric vector describing the trial in which the probabilistic reinforcement structure defined by the arguments probability, magnitude, and alternative applies.

Value

A data frame with columns 'probability', 'magnitude', 'alternative', and 'trial' describing the reinforcement structure of an individual RL Agent's arm.

Examples

#' # Define an arm that, when selected, leads to a 25% probability of receiving a
# positive reinforcement with magnitude 1 or zero for the first 75 trials and a negative
# reinforcement with magnitude 1 or zero with 75% probability for the second 75 trials.
rl_arms_define(
  probability = rep(0.25, 150),
  magnitude = c(rep(1, 75), rep(-1, 75)),
  alternative = 0,
  trial = 1:150
)
#>     probability magnitude alternative trial
#> 1          0.25         1           0     1
#> 2          0.25         1           0     2
#> 3          0.25         1           0     3
#> 4          0.25         1           0     4
#> 5          0.25         1           0     5
#> 6          0.25         1           0     6
#> 7          0.25         1           0     7
#> 8          0.25         1           0     8
#> 9          0.25         1           0     9
#> 10         0.25         1           0    10
#> 11         0.25         1           0    11
#> 12         0.25         1           0    12
#> 13         0.25         1           0    13
#> 14         0.25         1           0    14
#> 15         0.25         1           0    15
#> 16         0.25         1           0    16
#> 17         0.25         1           0    17
#> 18         0.25         1           0    18
#> 19         0.25         1           0    19
#> 20         0.25         1           0    20
#> 21         0.25         1           0    21
#> 22         0.25         1           0    22
#> 23         0.25         1           0    23
#> 24         0.25         1           0    24
#> 25         0.25         1           0    25
#> 26         0.25         1           0    26
#> 27         0.25         1           0    27
#> 28         0.25         1           0    28
#> 29         0.25         1           0    29
#> 30         0.25         1           0    30
#> 31         0.25         1           0    31
#> 32         0.25         1           0    32
#> 33         0.25         1           0    33
#> 34         0.25         1           0    34
#> 35         0.25         1           0    35
#> 36         0.25         1           0    36
#> 37         0.25         1           0    37
#> 38         0.25         1           0    38
#> 39         0.25         1           0    39
#> 40         0.25         1           0    40
#> 41         0.25         1           0    41
#> 42         0.25         1           0    42
#> 43         0.25         1           0    43
#> 44         0.25         1           0    44
#> 45         0.25         1           0    45
#> 46         0.25         1           0    46
#> 47         0.25         1           0    47
#> 48         0.25         1           0    48
#> 49         0.25         1           0    49
#> 50         0.25         1           0    50
#> 51         0.25         1           0    51
#> 52         0.25         1           0    52
#> 53         0.25         1           0    53
#> 54         0.25         1           0    54
#> 55         0.25         1           0    55
#> 56         0.25         1           0    56
#> 57         0.25         1           0    57
#> 58         0.25         1           0    58
#> 59         0.25         1           0    59
#> 60         0.25         1           0    60
#> 61         0.25         1           0    61
#> 62         0.25         1           0    62
#> 63         0.25         1           0    63
#> 64         0.25         1           0    64
#> 65         0.25         1           0    65
#> 66         0.25         1           0    66
#> 67         0.25         1           0    67
#> 68         0.25         1           0    68
#> 69         0.25         1           0    69
#> 70         0.25         1           0    70
#> 71         0.25         1           0    71
#> 72         0.25         1           0    72
#> 73         0.25         1           0    73
#> 74         0.25         1           0    74
#> 75         0.25         1           0    75
#> 76         0.25        -1           0    76
#> 77         0.25        -1           0    77
#> 78         0.25        -1           0    78
#> 79         0.25        -1           0    79
#> 80         0.25        -1           0    80
#> 81         0.25        -1           0    81
#> 82         0.25        -1           0    82
#> 83         0.25        -1           0    83
#> 84         0.25        -1           0    84
#> 85         0.25        -1           0    85
#> 86         0.25        -1           0    86
#> 87         0.25        -1           0    87
#> 88         0.25        -1           0    88
#> 89         0.25        -1           0    89
#> 90         0.25        -1           0    90
#> 91         0.25        -1           0    91
#> 92         0.25        -1           0    92
#> 93         0.25        -1           0    93
#> 94         0.25        -1           0    94
#> 95         0.25        -1           0    95
#> 96         0.25        -1           0    96
#> 97         0.25        -1           0    97
#> 98         0.25        -1           0    98
#> 99         0.25        -1           0    99
#> 100        0.25        -1           0   100
#> 101        0.25        -1           0   101
#> 102        0.25        -1           0   102
#> 103        0.25        -1           0   103
#> 104        0.25        -1           0   104
#> 105        0.25        -1           0   105
#> 106        0.25        -1           0   106
#> 107        0.25        -1           0   107
#> 108        0.25        -1           0   108
#> 109        0.25        -1           0   109
#> 110        0.25        -1           0   110
#> 111        0.25        -1           0   111
#> 112        0.25        -1           0   112
#> 113        0.25        -1           0   113
#> 114        0.25        -1           0   114
#> 115        0.25        -1           0   115
#> 116        0.25        -1           0   116
#> 117        0.25        -1           0   117
#> 118        0.25        -1           0   118
#> 119        0.25        -1           0   119
#> 120        0.25        -1           0   120
#> 121        0.25        -1           0   121
#> 122        0.25        -1           0   122
#> 123        0.25        -1           0   123
#> 124        0.25        -1           0   124
#> 125        0.25        -1           0   125
#> 126        0.25        -1           0   126
#> 127        0.25        -1           0   127
#> 128        0.25        -1           0   128
#> 129        0.25        -1           0   129
#> 130        0.25        -1           0   130
#> 131        0.25        -1           0   131
#> 132        0.25        -1           0   132
#> 133        0.25        -1           0   133
#> 134        0.25        -1           0   134
#> 135        0.25        -1           0   135
#> 136        0.25        -1           0   136
#> 137        0.25        -1           0   137
#> 138        0.25        -1           0   138
#> 139        0.25        -1           0   139
#> 140        0.25        -1           0   140
#> 141        0.25        -1           0   141
#> 142        0.25        -1           0   142
#> 143        0.25        -1           0   143
#> 144        0.25        -1           0   144
#> 145        0.25        -1           0   145
#> 146        0.25        -1           0   146
#> 147        0.25        -1           0   147
#> 148        0.25        -1           0   148
#> 149        0.25        -1           0   149
#> 150        0.25        -1           0   150