Initialize a Reinforcement Learning Agent

This is a generic function to create a new reinforcement learning agent.

rl_new_agent(model_type, model_id, ...)

Arguments

model_type

What type of RL Agent would you like to create? One of available_agents.

model_id

model_id (character) A model identifier referencing the reinforcement learning paradigm to perform the simulation.

...

Additional arguments to pass into the specific RL Agent. reinforcement learning paradigm to perform the simulation.

Value

A Reinforcement Learning Agent object of class model_type.

Examples

rl_new_agent(model_type = "tdrlConditioning",
             model_id = "Example TDRL",
             num_trials = 100,
             num_episodes = 10,
             gamma = 1,
             alpha = 0.2)
#> <tdrlConditioning Simulation Agent>
#> ID: "Example TDRL"
#>   100 Trials
#>   10 Episodes per Trial
#>   Temporal Discounting Factor of 1
#>   Learnrate of 0.2
#>   Reward Structure: ✖ not set
#>   Cue Structure: ✖ not set