Setup an agent's Reward Structure

Define the onset episode and offset episode of rewards for each trial

rl_set_reward(
  agent,
  reward_onset,
  reward_offset,
  reward_magnitude,
  keep_reward_structure = FALSE
)

Arguments

agent

An object of class "rlAgent" created with rl_new_agent.

reward_onset

(Numeric) The episode number a reward presentation begins. Either a single number where reward is presented at the same episode in all simulated trial or a numeric vector of length num_trials with reward onset occurring in different episodes.

reward_offset

(Numeric) The episode number a reward presentation ends. Either a single number where reward is presented at the same episode in all simulated trial or a numeric vector of length num_trials with reward offset occurring in different episodes.

reward_magnitude

(Numeric) The magnitude of the reward. Either a single number where reward is presented with the same magnitude across all simulated trials or a numeric vector of length num_trials with reward magnitude differing across trials.

keep_reward_structure

(Logical) FALSE (default) and any existing reward structure will be replaced when called. TRUE and the reward structure will be modified but will not remove previously defined rewards.

Value

The agent object with modified reward structure.